refactor: move utility functions to geometry-utils and remove unused util file
